use the correct XML in unit tests
The response in the DATA section of s3.t was of the wrong kind, so it did not get parsed properly and we ended up with an empty bucket list. We now use the correct response ListAllMyBucketsResult. I've also prepared this to include more responses so we can test the error handling in ::ResponseParser->_parse_error. A possible pitfall was that the example response at http://docs.aws.amazon.com/AmazonS3/latest/API/SOAPListAllMyBuckets.html not only has syntax errors, but also the namespace does not have a trailing slash which we are expecting. Since no-one is complaining and it seems to work in production I believe the documentation is wrong.

After the preparation for the tests of _parse_error I suspect that the expect_nothing check in the trigger in ::ResponseParser is the wrong way around. The ListAllMyBuckets request has a 0, which indicates not to expect nothing. That makes sense as the response is a list of buckets. But if it's false we don't check for errors. So we only check for errors if the response has no body. But because the result of the error check has been ignored since the first commit anyway this never made a difference.