add tests for ::FileIterator
There are a few things that are really hard to cover and I think they are not that pressing. The `$s->{prefix}` in `_fetch` is passed to as a paramter in the request to get a new page. It can be initialized in the constructor though Iterator::Paged, but is undocumented. It is used to limit bucket results. We should document it.